Se agrega una simulacion de batalla pokemon en el main, correciones en las clases de los pokemons ademas de las interfaces de ataques y se agregan comentarios al codigo para una mejor comprension

+++ src/Model/IAtaquesPokemonAgua.java
@@ -1,8 +1,9 @@
 package Model;
 
+//Interfaz para el manejo de los distintos ataques segun su tipo
 public interface IAtaquesPokemonAgua {
-    double pistolaAgua(int DefEspRivPokemon);
-    double hidroBomba(int DefEspRivPokemon);
+    double pistolaAgua(double DefEspRivPokemon);
+    double hidroBomba(double DefEspRivPokemon);
