+++ src/Model/IAtaquesPokemonTipoFuego.java
@@ -1,9 +1,10 @@
 package Model;
 
+//Interfaz para el manejo de los distintos ataques segun su tipo
 public interface IAtaquesPokemonTipoFuego {
 
-    double lanzallamas(int DefEspRivPokemon);
-    double llamarada(int DefEspRivPokemon);
+    double lanzallamas(double DefEspRivPokemon);
+    double llamarada(double DefEspRivPokemon);
